get participants from trip done
