+++ app/src/main/res/layout/contact_list_item.xml
@@ -79,14 +79,13 @@
 
     <Button
         android:id="@+id/btAdd"
-        android:visibility="gone"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginTop="42dp"
         android:layout_marginEnd="21dp"
         android:minHeight="30dp"
         android:background="@drawable/custom_button"
-        android:text="@string/add"
+        android:text="Ajouter"
         android:textAllCaps="false"
         android:textColor="@drawable/text_color_button"
         app:layout_constraintEnd_toEndOf="parent"
